api/requirements.txt
asyncpg==0.29.0
SQLAlchemy>=2.0
fastapi>=0.115
uvicorn[standard]>=0.30
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3
pydantic[email]>=2.8
python-dotenv>=1.0
httpx>=0.27

# Needed for OAuth2PasswordRequestForm (form-data)
python-multipart>=0.0.9

# Optional: enable /model router later
openai>=1.59.0
PyJWT==2.9.0
